+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 * Added `WechatPay.Plug.Refund` to handle refund notification from Wechat's Payment Gateway
+* Parse XML with `:xmerl`, drop dependency on `sweet_xml`.
 
 # v0.3.1